+++ webrepl_cli/cli.py
@@ -7,7 +7,8 @@ try:
     import usocket as socket
 except ImportError:
     import socket
-import websocket_helper
+
+from webrepl_cli import websocket_helper
 
 # Define to 1 to use builtin "uwebsocket" module of MicroPython
 USE_BUILTIN_UWEBSOCKET = 0
